+++ plugins/org.eclipse.actf.visualization.gui/src/org/eclipse/actf/accservice/swtbridge/internal/ia2/InternalAccessibleApplication.java
@@ -22,7 +22,7 @@ public class InternalAccessibleApplication implements AccessibleApplication {
 
     private IAccessibleApplication accessibleApplication = null;
     
-    public InternalAccessibleApplication(int address) {
+    public InternalAccessibleApplication(long address) {
         accessibleApplication = new IAccessibleApplication(address);
         accessibleApplication.AddRef();
     }
