+++ plugins/org.eclipse.actf.visualization.gui/src/org/eclipse/actf/accservice/swtbridge/internal/ia2/InternalAccessibleRelation.java
@@ -25,7 +25,7 @@ public class InternalAccessibleRelation implements AccessibleRelation {
 
     private IAccessibleRelation accessibleRelation = null;
     
-    public InternalAccessibleRelation(int address) {
+    public InternalAccessibleRelation(long address) {
         accessibleRelation = new IAccessibleRelation(address);
         accessibleRelation.AddRef();
     }
